# Skip flag processing for empty flags.