Speeds things up a bit, especially when there are a lot of empty flags and we can avoid some allocs.

```
benchmark                         old ns/op     new ns/op     delta
BenchmarkPrintEmptyParallel-8     14.2          14.1          -0.70%
BenchmarkFmtForComparison-8       6.17          6.16          -0.16%
BenchmarkCenteredParallel-8       122           122           +0.00%
BenchmarkLargeString-8            61284         40791         -33.44%
BenchmarkFmtLargeString-8         8011          8161          +1.87%
BenchmarkComplexFormat-8          257           257           +0.00%

benchmark                         old allocs     new allocs     delta
BenchmarkPrintEmptyParallel-8     0              0              +0.00%
BenchmarkFmtForComparison-8       0              0              +0.00%
BenchmarkCenteredParallel-8       3              3              +0.00%
BenchmarkLargeString-8            2002           1002           -49.95%
BenchmarkFmtLargeString-8         1              1              +0.00%
BenchmarkComplexFormat-8          13             13             +0.00%

benchmark                         old bytes     new bytes     delta
BenchmarkPrintEmptyParallel-8     0             0             +0.00%
BenchmarkFmtForComparison-8       0             0             +0.00%
BenchmarkCenteredParallel-8       132           132           +0.00%
BenchmarkLargeString-8            24144         20136         -16.60%
BenchmarkFmtLargeString-8         13583         13583         +0.00%
BenchmarkComplexFormat-8          344           344           +0.00%
```